Factor out random
