+++ bhv/cnative/run.sh
@@ -1,3 +1,3 @@
- g++ benchmark.cpp TurboSHAKEopt/TurboSHAKE.cpp TurboSHAKEopt/KeccakP-1600-opt64.cpp -O3 -std=c++20 -march=native -Wall -Wpedantic -Wextra
+ g++ testing.cpp TurboSHAKEopt/TurboSHAKE.cpp TurboSHAKEopt/KeccakP-1600-opt64.cpp -O3 -std=c++20 -march=native -Wall -Wpedantic -Wextra
  ./a.out
  rm a.out
